get footprints of search result from api and cli
The footprints for multiple scenes are returned as a geojson object
(api) or file (cli) and can be invoked from the command line with the
"-f" option.

The dictionary returned from get_product_info() now also includes the
footprint geometry as a string.

All tests, including for footprint, are passing.