Fix Unicode handling in requests
DHUS expects non-ASCII characters to be encoded in latin-1.
Added proper Unicode support by explicitly setting our encoding to UTF-8
in the request header.

Added a unit test for this as well, but VCR.py does not properly support
Unicode characters in request bodies, so I disabled this for now.