Pragmatic filename extension fix
This patch modifies the previous patch so that sentinelsat uses the
correct file extension for sentinel-5p files while not breaking half of
the tests. Note that the previous solution was the better one but would
unfortunately need quite a bit time for fixing test. If someone can
spare that time, that would be great… but for now, this fixes the issue.