Update download.py
The _download_with_node_filter function now uses the 'Filename' and 'manifest_name' dictionary keys to create the product folder and manifest filename depending on the type of product queried by the user.
The _dataobj_to_node_info function sets the "node_path" dictionary key using the path variable rather than recalling the "href" key of the dataobj_info object, so that the initial './' characters are removed if they were present in the original object. Not doing so causes the failure of the nodefilter function for Sentinel-3 products. Partial downloads of other products work normally.