Set the product folder path based on the product 'title'
Now the product folder path and the manifest filename are set based on the 'title' key of the odata dictionary. This ensures that Sentinel-3 paths and manifest filenames are set correctly, allowing nodefilters to function also with these products. The previously introduced 'Filename' key was removed in favour of the newly introduced 'manifest_name', which is then used to build the paths throughout the code.